LightFileIO: Splitting FileIO class into Reader and Writer classes
