+++ tests/python_functional/src/driver_io/file/tests/test_file_io.py
@@ -20,29 +20,33 @@
 # COPYING for details.
 #
 #############################################################################
-from src.driver_io.file.file_io import FileIO
+from src.driver_io.file.file_reader import FileReader
+from src.driver_io.file.file_writer import FileWriter
 
 
 def test_file_io_write_read(temp_file, test_message):
-    fileio = FileIO(temp_file)
-    fileio.write(test_message)
-    output = fileio.read()
+    file_writer = FileWriter(temp_file)
+    file_writer.write(test_message)
+    file_reader = FileReader(temp_file)
+    output = file_reader.read()
     assert test_message == output
 
 
 def test_file_io_multiple_write_read(temp_file, test_message):
-    fileio = FileIO(temp_file)
-    fileio.write(test_message)
-    fileio.write(test_message)
-    output = fileio.read()
+    file_writer = FileWriter(temp_file)
+    file_writer.write(test_message)
+    file_writer.write(test_message)
+    file_reader = FileReader(temp_file)
+    output = file_reader.read()
     assert test_message + test_message == output
 
 
 def test_file_io_rewrite_read(temp_file):
-    fileio = FileIO(temp_file)
+    file_writer = FileWriter(temp_file)
     content1 = "message 1\n"
-    fileio.write(content1)
-    assert fileio.read() == content1
+    file_writer.write(content1)
+    file_reader = FileReader(temp_file)
+    assert file_reader.read() == content1
     content2 = "message 2\n"
-    fileio.rewrite(content2)
-    assert fileio.read(position=0) == content2
+    file_writer.rewrite(content2)
+    assert file_reader.read(position=0) == content2
