+++ tests/python_functional/src/syslog_ng/console_log_reader.py
@@ -22,7 +22,7 @@
 #############################################################################
 import logging
 
-from src.driver_io.file.file_io import FileIO
+from src.driver_io.file.file_reader import FileReader
 from src.message_reader.message_reader import MessageReader
 from src.message_reader.message_reader import READ_ALL_AVAILABLE_MESSAGES
 from src.message_reader.single_line_parser import SingleLineParser
@@ -32,7 +32,7 @@ logger = logging.getLogger(__name__)
 
 class ConsoleLogReader(object):
     def __init__(self, instance_paths):
-        self.__stderr_io = FileIO(instance_paths.get_stderr_path())
+        self.__stderr_io = FileReader(instance_paths.get_stderr_path())
         self.__message_reader = MessageReader(self.__stderr_io.readline, SingleLineParser())
         self.__console_log_path = instance_paths.get_stderr_path()
